refactor: split GeometryHelper into individual geometry files
- Delete geometry.dart with GeometryHelper class
- Add plane.dart with quad, plane, and groundPlane geometries
- Add capsule.dart with capsule geometry
- Add cone.dart with conic geometry
- Add aabb.dart with fromAabb3 method
- Update utils.dart to export new geometry files
- Fix GeometryHelper references to use new split classes
