+++ thermion_dart/lib/src/utils/utils.dart
@@ -1,7 +1,16 @@
 library;
 
 export 'src/bone_data.dart';
-export 'src/geometry.dart';
+export 'src/geometry/aabb.dart';
+export 'src/geometry/camera.dart';
+export 'src/geometry/capsule.dart';
+export 'src/geometry/cone.dart';
+export 'src/geometry/cube.dart';
+export 'src/geometry/cylinder.dart';
+export 'src/geometry/plane.dart';
+export 'src/geometry/pyramid.dart';
+export 'src/geometry/quad.dart';
+export 'src/geometry/sphere.dart';
 export 'src/geometry/utils.dart';
 export 'src/gizmos.dart';
 export 'src/image.dart';
